+++ src/sound_opl.h
@@ -6,7 +6,7 @@ typedef struct opl_t
         int chip_nr[2];
         
         int64_t timers[2][2];
-        int timers_enable[2][2];
+        int64_t timers_enable[2][2];
 
         int16_t filtbuf[2];
